requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
# Testing dependencies (uncomment when ready to add tests)
# pytest==8.3.3
# pytest-asyncio==0.24.0
# Optional ML dependencies (uncomment based on your ML model needs)
# transformers==4.44.2
# torch==2.4.1
# httpx==0.27.2